[STR]Set STR_BUF to size_t to fix warnings with size_t

+++ src/libsac2c/support/str_buffer.h
@@ -5,7 +5,7 @@
 
 typedef struct STR_BUF str_buf;
 
-extern str_buf *SBUFcreate (int size);
+extern str_buf *SBUFcreate (size_t size);
 extern str_buf *SBUFprint (str_buf *s, const char *string);
 extern str_buf *SBUFprintf (str_buf *s, const char *format, ...);
 extern str_buf *SBUFvprintf (str_buf *s, const char *format, va_list arg_list);
@@ -13,7 +13,7 @@ extern char *SBUF2str (str_buf *s);
 extern void SBUFflush (str_buf *s);
 extern bool SBUFisEmpty (str_buf *s);
 extern str_buf *SBUFfree (str_buf *s);
-extern int SBUFlen (str_buf *s);
+extern size_t SBUFlen (str_buf *s);
 extern char *SBUFgetBuffer (str_buf *s);
 extern char *SBUF2strAndFree (str_buf **s);
